Extract rancher pages
