+++ tests/e2e/rancher/basepage.ts
