+++ tests/e2e/rancher/rancher-common.page.ts
