+++ tests/e2e/rancher/rancher-extensions.page.ts
